fix interestRate update field bug

+++ server/models/form.py
@@ -63,7 +63,7 @@ class UpdateFormModel(BaseModel):
     applicationType: Optional[str]
     loanAmount: Optional[float]
     loanTerm: Optional[int]
-    interestRate: Optional[int]
+    interestRate: Optional[float]
     companyName: Optional[str]
     coreBusinessType: Optional[str]
     staffNumber: Optional[int]
